+++ packages/std/src/math/decimal256.rs
@@ -42,6 +42,16 @@ impl Decimal256 {
 
     pub const MAX: Self = Self(Uint256::MAX);
 
+    /// Creates a Decimal from Uint256
+    pub const fn new(value: Uint256) -> Self {
+        Self(value)
+    }
+
+    /// Creates a Decimal from u128
+    pub const fn raw(value: u128) -> Self {
+        Self(Uint256::from_u128(value))
+    }
+
     /// Create a 1.0 Decimal256
     pub const fn one() -> Self {
         Self(Self::DECIMAL_FRACTIONAL)
@@ -490,6 +500,19 @@ mod tests {
     use super::*;
     use crate::errors::StdError;
     use crate::{from_slice, to_vec};
+
+    #[test]
+    fn decimal256_new() {
+        let expected = Uint256::from(300u128);
+        assert_eq!(expected, Decimal256::new(expected).0);
+    }
+
+    #[test]
+    fn decimal256_raw() {
+        let value = 300u128;
+        let expected = Uint256::from(value);
+        assert_eq!(expected, Decimal256::raw(value).0);
+    }
 
     #[test]
     fn decimal256_one() {