+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to
 - cosmwasm-std: Implement `MulAssign` for `Decimal`/`Decimal256`.
 - cosmwasm-std: Implement `is_zero`/`atomics`/`decimal_places` as const for Uint
   and Decimal types.
+- cosmwasm-std: Implement `new` and `raw` const constructors for
+  `Decimal`/`Decimal256`.
 - cosmwasm-crypto: Upgrade ed25519-zebra to version 3.
 
 ### Changed